move model classes to model package
