+++ app/src/main/java/com/sanaa/tudee_assistant/presentation/composable/TaskStatusTabs.kt
@@ -16,9 +16,9 @@ import com.sanaa.tudee_assistant.R
 import com.sanaa.tudee_assistant.presentation.designSystem.component.EmptyScreen
 import com.sanaa.tudee_assistant.presentation.designSystem.component.TabItem
 import com.sanaa.tudee_assistant.presentation.designSystem.component.TudeeScrollableTabs
+import com.sanaa.tudee_assistant.presentation.model.TaskUiState
 import com.sanaa.tudee_assistant.presentation.model.TaskUiStatus
 import com.sanaa.tudee_assistant.presentation.screen.tasks.TasksScreenUiState
-import com.sanaa.tudee_assistant.presentation.state.TaskUiState
 
 @Composable
 fun TaskStatusTabs(
