+++ app/src/main/java/com/sanaa/tudee_assistant/presentation/designSystem/component/CategoryItem.kt
@@ -22,7 +22,7 @@ import androidx.compose.ui.unit.dp
 import com.sanaa.tudee_assistant.presentation.composable.CategoryThumbnail
 import com.sanaa.tudee_assistant.presentation.designSystem.theme.Theme
 import com.sanaa.tudee_assistant.presentation.designSystem.theme.TudeeTheme
-import com.sanaa.tudee_assistant.presentation.state.CategoryUiState
+import com.sanaa.tudee_assistant.presentation.model.CategoryUiState
 import com.sanaa.tudee_assistant.presentation.utils.DataProvider
 
 @Composable
