+++ app/src/main/java/com/sanaa/tudee_assistant/presentation/model/mapper/TaskStateMapper.kt
@@ -1,12 +1,12 @@
-package com.sanaa.tudee_assistant.presentation.state.mapper
+package com.sanaa.tudee_assistant.presentation.model.mapper
 
 import com.sanaa.tudee_assistant.domain.model.AddTaskRequest
 import com.sanaa.tudee_assistant.domain.model.Task
 import com.sanaa.tudee_assistant.domain.model.Task.TaskPriority
 import com.sanaa.tudee_assistant.domain.model.Task.TaskStatus
 import com.sanaa.tudee_assistant.presentation.model.TaskUiPriority
+import com.sanaa.tudee_assistant.presentation.model.TaskUiState
 import com.sanaa.tudee_assistant.presentation.model.TaskUiStatus
-import com.sanaa.tudee_assistant.presentation.state.TaskUiState
 import kotlinx.datetime.LocalDate
 import kotlinx.datetime.LocalDateTime
 
@@ -43,7 +43,6 @@ fun TaskStatus.toState(): TaskUiStatus {
         TaskStatus.DONE -> TaskUiStatus.DONE
     }
 }
-
 
 fun TaskUiState.toTask(): Task {
     return Task(
