+++ app/src/main/java/com/sanaa/tudee_assistant/presentation/screen/category/CategoryInteractionListener.kt
@@ -1,7 +1,7 @@
 package com.sanaa.tudee_assistant.presentation.screen.category
 
 import android.net.Uri
-import com.sanaa.tudee_assistant.presentation.state.CategoryUiState
+import com.sanaa.tudee_assistant.presentation.model.CategoryUiState
 
 interface CategoryInteractionListener {
     fun onAddCategory(newCategory: CategoryUiState)
