+++ app/src/main/java/com/sanaa/tudee_assistant/presentation/screen/categoryTask/CategoryTaskInteractionListener.kt
@@ -1,8 +1,8 @@
 package com.sanaa.tudee_assistant.presentation.screen.categoryTask
 
 import android.net.Uri
-import com.sanaa.tudee_assistant.presentation.state.CategoryUiState
-import com.sanaa.tudee_assistant.presentation.state.TaskUiState
+import com.sanaa.tudee_assistant.presentation.model.CategoryUiState
+import com.sanaa.tudee_assistant.presentation.model.TaskUiState
 
 interface CategoryTaskInteractionListener {
     fun onDeleteClicked()
