+++ app/src/main/java/com/sanaa/tudee_assistant/presentation/screen/categoryTask/CategoryTaskScreenUiState.kt
@@ -1,8 +1,8 @@
 package com.sanaa.tudee_assistant.presentation.screen.categoryTask
 
+import com.sanaa.tudee_assistant.presentation.model.CategoryUiState
+import com.sanaa.tudee_assistant.presentation.model.TaskUiState
 import com.sanaa.tudee_assistant.presentation.model.TaskUiStatus
-import com.sanaa.tudee_assistant.presentation.state.CategoryUiState
-import com.sanaa.tudee_assistant.presentation.state.TaskUiState
 
 data class CategoryTaskScreenUiState(
     val isLoading: Boolean = false,
